moved the apikey to buildConfigField on the gradle file and added the error option
