+++ app/src/main/java/com/example/weatherapp/data/service/WeatherService.kt
@@ -1,5 +1,6 @@
 package com.example.weatherapp.data.service
 
+import com.example.weatherapp.BuildConfig
 import com.example.weatherapp.data.service.api.WeatherApi
 import com.example.weatherapp.data.service.request.generator.WeatherRequestGenerator
 import com.example.weatherapp.data.service.response.WeatherResponse
@@ -11,7 +12,7 @@ class WeatherService {
 
     fun getWeatherByCityName(city: String): Observable<WeatherResponse> {
         return Observable.create { subscriber ->
-            val callResponse = api.createService(WeatherApi::class.java).getWeatherData(city, API_KEY, METRIC)
+            val callResponse = api.createService(WeatherApi::class.java).getWeatherData(city, BuildConfig.API_KEY, METRIC)
             val response = callResponse.execute()
             if (response.isSuccessful) {
                 response.body()?.let { subscriber.onNext(it) }
@@ -22,7 +23,6 @@ class WeatherService {
     }
 
     companion object {
-        private const val API_KEY = "027b7152f20f23f8f0cbc9f4bfab880c"
         private const val METRIC = "metric"
     }
 }
