+++ app/src/main/java/com/example/weatherapp/mvp/contract/WeatherContract.kt
@@ -1,4 +1,4 @@
-package com.example.weatherapp.mvp
+package com.example.weatherapp.mvp.contract
 
 import com.example.weatherapp.data.service.response.WeatherResponse
 import io.reactivex.rxjava3.core.Observable
@@ -15,5 +15,6 @@ interface WeatherContract {
 
     interface View {
         fun showData(data: WeatherResponse)
+        fun showError()
     }
 }
