+++ app/src/main/java/com/example/weatherapp/mvp/model/WeatherModel.kt
@@ -2,7 +2,7 @@ package com.example.weatherapp.mvp.model
 
 import com.example.weatherapp.data.service.WeatherService
 import com.example.weatherapp.data.service.response.WeatherResponse
-import com.example.weatherapp.mvp.WeatherContract
+import com.example.weatherapp.mvp.contract.WeatherContract
 import io.reactivex.rxjava3.core.Observable
 
 class WeatherModel(private val weatherService: WeatherService) : WeatherContract.Model {
