+++ gradle.properties
@@ -19,3 +19,5 @@ android.useAndroidX=true
 android.enableJetifier=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
+# API Key
+API_KEY="027b7152f20f23f8f0cbc9f4bfab880c"
